requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3
mediapipe==0.10.11
pyannote.audio==3.1.1
torch>=2.0.0
torchaudio>=2.0.0
opencv-python==4.9.0.80
numpy>=1.24.0
scipy>=1.12.0
ffmpeg-python==0.2.0
